requirements.txt
discord.py==2.3.2
huggingface_hub==0.20.3
python-dotenv